加入者情報を削除する際、指定されたIDが存在するか確認するようにした

+++ src/main/java/com/s_giken/training/webapp/controller/MemberController.java
@@ -127,6 +127,11 @@ public class MemberController {
 	public String deleteMember(
 			@PathVariable int id,
 			RedirectAttributes redirectAttributes) {
+		var member = memberService.findById(id);
+		if (!member.isPresent()) {
+			throw new NotFoundException(String.format("指定したmemberId(%d)の加入者情報が存在しません。", id));
+		}
+
 		memberService.deleteById(id);
 		redirectAttributes.addFlashAttribute("message", "削除しました。");
 		return "redirect:/member/search";
